requirements.txt
fastapi
uvicorn
httpx
passlib
python-jose[cryptography]
sqlalchemy[asyncio]
aiosqlite
pytest